add ModelConfig to step deco
