+++ src/zenml/model/base_model.py
@@ -1,0 +1,61 @@
+#  Copyright (c) ZenML GmbH 2023. All Rights Reserved.
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at:
+#
+#       https://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+#  or implied. See the License for the specific language governing
+#  permissions and limitations under the License.
+"""Model base model to support Model WatchTower feature."""
+
+from typing import List, Optional
+
+from pydantic import BaseModel, Field
+
+STR_FIELD_MAX_LENGTH = 255
+TEXT_FIELD_MAX_LENGTH = 65535
+
+
+class ModelBaseModel(BaseModel):
+    """Model base model."""
+
+    name: str = Field(
+        title="The name of the model",
+        max_length=STR_FIELD_MAX_LENGTH,
+    )
+    license: Optional[str] = Field(
+        title="The license model created under",
+        max_length=TEXT_FIELD_MAX_LENGTH,
+    )
+    description: Optional[str] = Field(
+        title="The description of the model",
+        max_length=TEXT_FIELD_MAX_LENGTH,
+    )
+    audience: Optional[str] = Field(
+        title="The target audience of the model",
+        max_length=TEXT_FIELD_MAX_LENGTH,
+    )
+    use_cases: Optional[str] = Field(
+        title="The use cases of the model",
+        max_length=TEXT_FIELD_MAX_LENGTH,
+    )
+    limitations: Optional[str] = Field(
+        title="The know limitations of the model",
+        max_length=TEXT_FIELD_MAX_LENGTH,
+    )
+    trade_offs: Optional[str] = Field(
+        title="The trade offs of the model",
+        max_length=TEXT_FIELD_MAX_LENGTH,
+    )
+    ethic: Optional[str] = Field(
+        title="The ethical implications of the model",
+        max_length=TEXT_FIELD_MAX_LENGTH,
+    )
+    tags: Optional[List[str]] = Field(
+        title="Tags associated with the model",
+    )
